Introducing new parent class for developer/team (company) app storages.
